split Simulator and SimulatorDialog into separate files
